+++ pnet-data-api-java/src/main/java/pnet/data/api/company/CompanyDataDTO.java
@@ -72,18 +72,18 @@ public class CompanyDataDTO implements WithCompanyId, WithMatchcode, WithTenants
     @Schema(description = "Valid tenants of the company (also known as Portal-ID).")
     private Collection<String> tenants;
 
-    @Schema(description =
-        "All brands assigned to the company. The matchcode of each item fits to the matchcodes of the brands "
+    @Schema(
+        description = "All brands assigned to the company. The matchcode of each item fits to the matchcodes of the brands "
             + "interface.")
     private Collection<CompanyBrandLinkDTO> brands;
 
-    @Schema(description =
-        "All contract types assigned to the company. The matchcode of each item fits to the matchcodes of the "
+    @Schema(
+        description = "All contract types assigned to the company. The matchcode of each item fits to the matchcodes of the "
             + "contract types interface.")
     private Collection<CompanyContractTypeLinkDTO> contractTypes;
 
-    @Schema(description =
-        "All contract states assigned to the company. The matchcode of each item fits to the matchcodes of the "
+    @Schema(
+        description = "All contract states assigned to the company. The matchcode of each item fits to the matchcodes of the "
             + "contract states interface.")
     private Collection<CompanyContractStateLinkDTO> contractStates;
 
@@ -111,8 +111,8 @@ public class CompanyDataDTO implements WithCompanyId, WithMatchcode, WithTenants
     @Schema(description = "If true, the company is partly configured and managed by the BPCM system.")
     private boolean bpcmManaged;
 
-    @Schema(description =
-        "All additional numbers of the company. The matchcode of each item fits to the matchcodes of the "
+    @Schema(
+        description = "All additional numbers of the company. The matchcode of each item fits to the matchcodes of the "
             + "company number types interface.")
     private Collection<CompanyNumberLinkDTO> additionalNumbers;
 
@@ -188,6 +188,12 @@ public class CompanyDataDTO implements WithCompanyId, WithMatchcode, WithTenants
     @Schema(description = "A link to the Telegam page of the company.")
     private String telegramLink;
 
+    @Schema(description = "A link to the TikTok page of the company.")
+    private String tikTokLink;
+
+    @Schema(description = "A link to the X(Twitter) page of the company.")
+    private String xTwitterLink;
+
     @Schema(description = "The matchcode of the legal form.")
     private String legalFormMatchcode;
 
@@ -206,8 +212,8 @@ public class CompanyDataDTO implements WithCompanyId, WithMatchcode, WithTenants
     @Schema(description = "The venue jurisdiction the company is part of because of its geographical position.")
     private String jurisdiction;
 
-    @Schema(description =
-        "The provision in a company's constitution stating the purpose and range of activities for which the "
+    @Schema(
+        description = "The provision in a company's constitution stating the purpose and range of activities for which the "
             + "company is carried on (part of the impressum).")
     private String objectsClause;
 
@@ -419,9 +425,8 @@ public class CompanyDataDTO implements WithCompanyId, WithMatchcode, WithTenants
     public Optional<CompanyContractDistributionStructureLinkDTO> findContractDistributionStructure(
         Predicate<? super CompanyContractDistributionStructureLinkDTO> predicate)
     {
-        return contractDistributionStructure == null ?
-            Optional.empty() :
-            contractDistributionStructure.stream().filter(predicate).findFirst();
+        return contractDistributionStructure == null ? Optional.empty()
+            : contractDistributionStructure.stream().filter(predicate).findFirst();
     }
 
     public void setContractDistributionStructure(
@@ -745,6 +750,26 @@ public class CompanyDataDTO implements WithCompanyId, WithMatchcode, WithTenants
         this.telegramLink = telegramLink;
     }
 
+    public String getTikTokLink()
+    {
+        return tikTokLink;
+    }
+
+    public void setTikTokLink(String tikTokLink)
+    {
+        this.tikTokLink = tikTokLink;
+    }
+
+    public String getXTwitterLink()
+    {
+        return xTwitterLink;
+    }
+
+    public void setXTwitterLink(String xTwitterLink)
+    {
+        this.xTwitterLink = xTwitterLink;
+    }
+
     public String getLegalFormMatchcode()
     {
         return legalFormMatchcode;
@@ -929,9 +954,8 @@ public class CompanyDataDTO implements WithCompanyId, WithMatchcode, WithTenants
     public Optional<CompanyAdvisorAssignmentLinkDTO> findAdvisorAssignment(
         Predicate<? super CompanyAdvisorAssignmentLinkDTO> predicate)
     {
-        return advisorAssignments == null ?
-            Optional.empty() :
-            advisorAssignments.stream().filter(predicate).findFirst();
+        return advisorAssignments == null ? Optional.empty()
+            : advisorAssignments.stream().filter(predicate).findFirst();
     }
 
     public void setAdvisorAssignments(Collection<CompanyAdvisorAssignmentLinkDTO> advisorAssignments)
@@ -952,27 +976,29 @@ public class CompanyDataDTO implements WithCompanyId, WithMatchcode, WithTenants
     @Override
     public String toString()
     {
-        return String.format(
-            "CompanyDataDTO [companyId=%s, matchcode=%s, administrativeTenant=%s, label=%s, name=%s, nameAffix=%s, "
-                + "additionalNameAffix=%s, marketingName=%s, groupMembers=%s, tenants=%s, brands=%s, contractTypes=%s, "
-                + "contractStates=%s, contractDistributionStructure=%s, vatIdNumber=%s, sapNumber=%s, "
-                + "companyNumber=%s, bpcmLocationUuid=%s, bpcmManaged=%s, additionalNumbers=%s, street=%s, city=%s, "
-                + "postalCode=%s, countryCode=%s, country=%s, region=%s, iban=%s, bic=%s, types=%s, phoneNumber=%s, "
-                + "mobileNumber=%s, speedDial=%s, faxNumber=%s, email=%s, leadEmail=%s, homepage=%s, postal=%s, "
-                + "facebookLink=%s, youTubeLink=%s, instagramLink=%s, viberLink=%s, telegramLink=%s, "
-                + "legalFormMatchcode=%s, dataProcessingRegisterNumber=%s, commercialRegisterNumber=%s, "
-                + "certificateType=%s, certificateNumber=%s, jurisdiction=%s, objectsClause=%s, generalPartner=%s, "
-                + "registeredOffice=%s, chamberAffiliation=%s, commercialRegulations=%s, regulatoryAuthority=%s, "
-                + "arbitrationBoard=%s, additionalImprintInfo=%s, businessInformationNumber=%s, location=%s, "
-                + "externalBrands=%s, advisorAssignments=%s, lastUpdate=%s]", companyId, matchcode,
-            administrativeTenant, label, name, nameAffix, additionalNameAffix, marketingName, groupMembers, tenants,
-            brands, contractTypes, contractStates, contractDistributionStructure, vatIdNumber, sapNumber, companyNumber,
-            bpcmLocationUuid, bpcmManaged, additionalNumbers, street, city, postalCode, countryCode, country, region,
-            iban, bic, types, phoneNumber, mobileNumber, speedDial, faxNumber, email, leadEmail, homepage, postal,
-            facebookLink, youTubeLink, instagramLink, viberLink, telegramLink, legalFormMatchcode,
-            dataProcessingRegisterNumber, commercialRegisterNumber, certificateType, certificateNumber, jurisdiction,
-            objectsClause, generalPartner, registeredOffice, chamberAffiliation, commercialRegulations,
-            regulatoryAuthority, arbitrationBoard, additionalImprintInfo, businessInformationNumber, location,
-            externalBrands, advisorAssignments, lastUpdate);
+        return String
+            .format(
+                "CompanyDataDTO [companyId=%s, matchcode=%s, administrativeTenant=%s, label=%s, name=%s, nameAffix=%s, "
+                    + "additionalNameAffix=%s, marketingName=%s, groupMembers=%s, tenants=%s, brands=%s, contractTypes=%s, "
+                    + "contractStates=%s, contractDistributionStructure=%s, vatIdNumber=%s, sapNumber=%s, "
+                    + "companyNumber=%s, bpcmLocationUuid=%s, bpcmManaged=%s, additionalNumbers=%s, street=%s, city=%s, "
+                    + "postalCode=%s, countryCode=%s, country=%s, region=%s, iban=%s, bic=%s, types=%s, phoneNumber=%s, "
+                    + "mobileNumber=%s, speedDial=%s, faxNumber=%s, email=%s, leadEmail=%s, homepage=%s, postal=%s, "
+                    + "facebookLink=%s, youTubeLink=%s, instagramLink=%s, viberLink=%s, telegramLink=%s, tikTokLink=%s, "
+                    + "xTwitterLink=%s, legalFormMatchcode=%s, dataProcessingRegisterNumber=%s, commercialRegisterNumber=%s, "
+                    + "certificateType=%s, certificateNumber=%s, jurisdiction=%s, objectsClause=%s, generalPartner=%s, "
+                    + "registeredOffice=%s, chamberAffiliation=%s, commercialRegulations=%s, regulatoryAuthority=%s, "
+                    + "arbitrationBoard=%s, additionalImprintInfo=%s, businessInformationNumber=%s, location=%s, "
+                    + "externalBrands=%s, advisorAssignments=%s, lastUpdate=%s]",
+                companyId, matchcode, administrativeTenant, label, name, nameAffix, additionalNameAffix, marketingName,
+                groupMembers, tenants, brands, contractTypes, contractStates, contractDistributionStructure,
+                vatIdNumber, sapNumber, companyNumber, bpcmLocationUuid, bpcmManaged, additionalNumbers, street, city,
+                postalCode, countryCode, country, region, iban, bic, types, phoneNumber, mobileNumber, speedDial,
+                faxNumber, email, leadEmail, homepage, postal, facebookLink, youTubeLink, instagramLink, viberLink,
+                telegramLink, tikTokLink, xTwitterLink, legalFormMatchcode, dataProcessingRegisterNumber,
+                commercialRegisterNumber, certificateType, certificateNumber, jurisdiction, objectsClause,
+                generalPartner, registeredOffice, chamberAffiliation, commercialRegulations, regulatoryAuthority,
+                arbitrationBoard, additionalImprintInfo, businessInformationNumber, location, externalBrands,
+                advisorAssignments, lastUpdate);
     }
 }